Backend/requirements.txt
# FastAPI 
fastapi==0.95.0
uvicorn==0.21.0
python-multipart==0.0.5
pydantic==1.10.7
jinja2==3.1.2

# Django and REST framework
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.0.0
whitenoise==6.4.0
dj-database-url==1.0.0
google-cloud-storage

# Authentication and Security
PyJWT==2.6.0
cryptography==40.0.0
twilio==7.0.0  
six==1.16.0

# Celery 
celery==5.1.2
redis==3.5.3
flower==1.0.0

# Database
psycopg2-binary==2.9.7  

# Email
django-anymail==10.0  

# Data processing and analysis
numpy==1.21.2
pandas==1.3.3

# OCR and image processing
Pillow==8.3.2  
pytesseract==0.3.8
opencv-python-headless==4.5.3.56
pdf2image==1.16.0
PyMuPDF==1.18.14

# NLP and text processing
spacy==3.1.3
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0.tar.gz
dateparser==1.0.0
price-parser==0.3.4
usaddress==0.5.10
pycountry==20.7.3

# Machine learning and deep learning
torch==1.9.1
transformers==4.11.3

# Google Cloud Vision and Document AI
google-cloud-documentai==2.15.0
google-cloud-vision==3.4.4

# File handling and compression
python-magic==0.4.24
zipfile36==0.1.3

# Excel handling
openpyxl==3.0.9
XlsxWriter==3.0.2

# Asynchronous programming
aiohttp==3.7.4.post0
async-timeout==3.0.1
aiofiles==0.8.0
aioredis==2.0.1

# System utilities
psutil==5.8.0

# API Documentation
drf-yasg==1.21.5

# Filtering
django-filter==23.2

# Error handling and retries
tenacity==8.0.1

# Rate limiting
ratelimit==2.2.1

# Logging and error tracking
sentry-sdk==1.4.3

# Utilities
python-dotenv==1.0.0
requests==2.28.0

# Testing
pytest==6.2.5  
pytest-asyncio==0.15.1

# Development tools
black==21.9b0  
flake8==3.9.2  
isort==5.9.3

# Production deployment
gunicorn==20.1.0
